# docs/openapi/journey-simulator.yaml
openapi: 3.0.3
info:
  title: Money Factory AI — Journey Simulator API
  version: 0.1.0
  description: >-
    Public API for Journey Simulator (frontend Next.js + backend routes). Includes
    health, journeys, step/submit/state, agents logs, Solana devnet utilities
    (mint, tx), RAG stubs, SIWS auth stubs, and DAO/Stake simulations.
servers:
  - url: https://journey.mfai.app
    description: Production
  - url: http://localhost:3000
    description: Local dev
...
paths:
...
  /api/mint/execute:
    post:
      tags: [Mint]
      summary: Execute mint (devnet)
      parameters:
        - in: header
          name: x-user-id
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sim:
                  $ref: '#/components/schemas/MintSimulationInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  tx:
                    type: object
                    properties:
                      txSig: { type: string }
        '400': { description: Missing/invalid input or minter key missing }
        '403': { description: Kill switch enabled }
        '500': { description: Execution failed }
...
components:
  schemas:
...
    MintSimulationInput:
      type: object
      required: [ok, estFeeLamports, riskScore, network]
      properties:
        ok: { type: boolean }
        estFeeLamports: { type: number }
        riskScore: { type: number }
        txB64: { type: string }
        network: { type: string }
...